wip: this doesn't work cuz of cgo problems hopefully go 1.8 will fix
Update consul to 0.7.5

Additionally, update all shared dependencies with consul for which
0.7.5 consul has a newer vendored revision than we do.

This commit brings in a lot of new packages, however our main dependency
growth is due to a dependence on
github.com/hashicorp/consul/command/agent which is only used within test
files. As a result the increase in vendored dependencies should not
actually increase binary size.